insurance_extractor/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0,<3
python-docx>=1.1,<2
pytest>=8.0,<9
